Component personTail moved from features folder
