+++ src/components/personTail/styled.js
